:wrench:

+++ packages/skia/android/src/main/java/com/shopify/reactnative/skia/SkiaBaseView.java
@@ -81,5 +81,5 @@ public abstract class SkiaBaseView extends ReactViewGroup implements SkiaViewAPI
 
     protected abstract void unregisterView();
 
-    protected abstract byte[] getBitmap(int width, int height);
+    protected abstract int[] getBitmap(int width, int height);
 }
